Refactor ward management to use NextJS dynamic routes
