+++ pageTests/trust-admin/edit-a-ward-success.test.js
@@ -1,4 +1,4 @@
-import { getServerSideProps } from "../../pages/trust-admin/edit-a-ward-success";
+import { getServerSideProps } from "../../pages/trust-admin/wards/[id]/edit-success";
 
 const authenticatedReq = {
   headers: {
@@ -6,7 +6,7 @@ const authenticatedReq = {
   },
 };
 
-describe("/trust-admin/edit-a-ward-success", () => {
+describe("/trust-admin/wards/[id]/edit-success", () => {
   const anonymousReq = {
     headers: {
       cookie: "",
@@ -54,7 +54,7 @@ describe("/trust-admin/edit-a-ward-success", () => {
           req: authenticatedReq,
           res,
           query: {
-            wardId: "ward ID",
+            id: "ward ID",
           },
           container,
         });
@@ -81,7 +81,7 @@ describe("/trust-admin/edit-a-ward-success", () => {
           req: authenticatedReq,
           res,
           query: {
-            wardId: "1",
+            id: "1",
           },
           container,
         });
